refactor: remove sections from partial views. Pass `rankingApiParam` to `LaNationalRankChart` component. Define const for copy in index
