+++ front-end-components/src/views/la-national-rank/index.tsx
@@ -1,3 +1,13 @@
 /* eslint-disable react-refresh/only-export-components */
 export * from "src/views/la-national-rank/types";
 export * from "src/views/la-national-rank/view";
+
+export const plannedExpenditureTitle = "planned-expenditure";
+export const plannedExpenditureSummary = "Outturn figures are pre-recoupment";
+export const plannedExpenditureValueLabel =
+  "Spend as percentage of planned expenditure";
+
+export const fundingTitle = "funding";
+export const fundingSummary =
+  "DSG Funding and outturn are pre-recoupment figures. Outturn more than 100% indicates that spend is greater than the funding allocation.";
+export const fundingValueLabel = "Spend as percentage of funding";
