+++ front-end-components/src/views/la-national-rank/partials/index.tsx
@@ -1,5 +1,3 @@
 /* eslint-disable react-refresh/only-export-components */
 export * from "src/views/la-national-rank/partials/types";
 export * from "src/views/la-national-rank/partials/chart";
-export * from "src/views/la-national-rank/partials/funding-section";
-export * from "src/views/la-national-rank/partials/planned-expenditure-section";
